prefactor: extract to useTableSorting helper
